services/processor-service/pyproject.toml
[project]
name = "processor-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "httpx>=0.28.1",
    "loadenv>=0.1.1",
    "pika>=1.3.2",
    "python-telegram-bot>=22.5",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
    "telebot>=0.0.5",
    "telethon>=1.42.0",
]
